Extract post summary to own component
