+++ src/components/PostList.css
@@ -50,12 +50,3 @@
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
 }
-
-.post-summary {
-  font-size: 12px;
-}
-
-.post-author,
-.post-timestamp {
-  font-weight: bold;
-}
