+++ src/components/PostSummary.css
@@ -1,0 +1,8 @@
+.post-summary {
+  font-size: 12px;
+}
+
+.post-author,
+.post-timestamp {
+  font-weight: bold;
+}
